# stripstreamlib-openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "api", "description": "", "license": { "name": "MIT" }, "version": "3.32.0" },
  "paths": {
...
    "/books/{id}/thumbnail": {
      "get": {
        "tags": ["books"],
        "summary": "Get book thumbnail image",
        "operationId": "get_thumbnail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Book UUID",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "WebP thumbnail image" },
          "304": { "description": "Not modified (matches If-None-Match)" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Book not found or thumbnail not available" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your API Bearer token (format: stl_<prefix>_<secret>)"
      }
    }
  },
  "security": [{ "Bearer": [] }],
...
}